# frontend/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"Backend API - Order Management","description":"REST API to manage customer orders, order lines and attachments (SQLite)","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/orders/{orderId}":{"get":{"tags":["Orders"],"operationId":"getOrderById","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Order"}}}}}},"put":{"tags":["Orders"],"operationId":"updateOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Order"}}}}}},"delete":{"tags":["Orders"],"operationId":"deleteOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/orders":{"get":{"tags":["Orders"],"operationId":"getAllOrders","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Order"}}}}}}},"post":{"tags":["Orders"],"operationId":"createOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Order"}}}}}}},"/api/orders/{orderId}/lines":{"get":{"tags":["Order Lines"],"operationId":"getAllOrderLines","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderLine"}}}}}}},"post":{"tags":["Order Lines"],"operationId":"createOrderLine","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderLine"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderLine"}}}}}}},"/api/orders/{orderId}/attachemets":{"get":{"tags":["Attachments"],"operationId":"getAllAttachments","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Attachment"}}}}}}},"post":{"tags":["Attachments"],"operationId":"createAttachment","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Attachment"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Attachment"}}}}}}},"/api/orders/{orderId}/lines/{lineId}":{"get":{"tags":["Order Lines"],"operationId":"getOrderLineById","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"lineId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderLine"}}}}}}},"/api/orders/{orderId}/attachemets/{attachemetsid}":{"get":{"tags":["Attachments"],"operationId":"getAttachmentById","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"attachemetsid","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Attachment"}}}}}},"delete":{"tags":["Attachments"],"operationId":"deleteAttachment","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"attachemetsid","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/orders/{orderId}/lines/{orderLineId}":{"delete":{"tags":["Order Lines"],"operationId":"deleteOrderLine","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"orderLineId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Attachment":{"type":"object","properties":{"id":{"type":"integer","format":"int32","readOnly":true},"typeFile":{"type":"string","example":"pdf"},"urlFile":{"type":"string","example":"http://files.example.com/invoice.pdf"}},"required":["typeFile","urlFile"]},"Order":{"type":"object","properties":{"id":{"type":"integer","format":"int32","readOnly":true},"dateOrder":{"type":"string","format":"date-time","example":"2026-06-20T10:00:00"},"status":{"type":"string","example":"PENDING","maxLength":25,"minLength":0},"orderLines":{"type":"array","items":{"$ref":"#/components/schemas/OrderLine"},"readOnly":true},"attachments":{"type":"array","items":{"$ref":"#/components/schemas/Attachment"},"readOnly":true}},"required":["dateOrder"]},"OrderLine":{"type":"object","properties":{"id":{"type":"integer","format":"int32","readOnly":true},"designation":{"type":"string","example":"Laptop","maxLength":50,"minLength":0},"quantity":{"type":"integer","format":"int32","example":1},"price":{"type":"number","format":"double","example":999.99}},"required":["designation","price","quantity"]}}}}